PrepMind/PrepMind/requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-groq
faiss-cpu
fastembed
pypdf
beautifulsoup4
requests
python-dotenv
